requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
onnxslim==0.1.35
opencv-python==4.10.0.84
openvino==2024.4.0
openvino-dev==2024.4.0
openvino-telemetry==2024.1.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.2.0
six==1.16.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.0
ultralytics-thop==2.0.9
urllib3==2.2.3
